Create Bulk Reset Passwords in Active Directory
Note: This script assumes that you have the appropriate permissions to reset passwords in Active Directory and that the users exist in Active Directory. Also, please make sure to use a strong and complex password for the users.
It's also recommended to store the usernames in a csv file and read from it instead of hardcoding in the script.