Create Check Software Version on Multiple Clients
In this script, the list of computers is defined at the beginning of the script, and the list of software to check is also defined. The script then loops through each computer and checks if it is online using the Test-Connection cmdlet. If the computer is online, the script loops through the list of software and uses the Get-WmiObject cmdlet to get the version of the software installed on the computer. If the software is found to be installed, the script writes the version to the console, otherwise it shows that the software is not installed on that computer.

You can also add or remove software from the list, and add or remove computers from the list as well.

Please note that this script uses WMI (Windows Management Instrumentation) to retrieve the software version, so it will only work for software that is installed using Windows Installer. Some software may not show the version correctly or at all.