Create Schedule Computer to Shutdown Daily
This script schedules a computer to shut down at the time specified in the $shutdownTime variable, which is in the format of "HH:mm:ss". The script then retrieves the current time, and compares it to the shutdown time. If the current time is later than the shutdown time, it schedules the shutdown for the next day. The script then uses the "schtasks" command to create a new task that runs the "shutdown /s /t 0" command at the scheduled time, which will shut down the computer.

You can customize the time by changing the $shutdownTime variable.

Please note that you need to run this script as an Administrator in order to create the scheduled task. Also, you need to run it on the computer you want to schedule the shutdown on.

Keep in mind that shutting down a computer forcefully may cause unsaved data loss and other issues, so it is recommended to save all your work and close all the running applications before running this script.