Create Bulk Update Software
This script updates the software specified in the $software variable on the computers specified in the $computers variable. The script uses the Test-Connection cmdlet to check if the computer is online, and if it is, it uses the Get-WmiObject cmdlet to check if the software is installed on the computer. If the software is installed, the script uses the Invoke